imageRepository specified to reduce bandwidth use
To replicate this setup, first arrange for a registry to be hosted on
your local network, if possible.

--git-url is also set here, replace:
git@github.com:kingdonb/wks-quickstart-firekube.git
with your own repo URL and ensure that cluster-key remains protected
whenever deploying any of this anywhere adjacent to a public network.

The cost of bandwidth is an issue in terms of both time and financial,
so a localized registry should be used in every case whenever possible.

Use fast local networks, even same-host cache or proxy when possible. In
this case, our host uses Docker which has been configured to talk to an
insecure image repo hosted from /tmp, at an IP on the same host bridge.

(If using Docker for Mac, you may need to modify your host machine's
Docker daemon config in order to use an insecure registry unless it is
made available on localhost:5000 -- docker-config.yaml here includes an
example daemon.json with setting for "insecure-registries" configured to
allow several insecure registries that are present on my local network.)